Time and Date Done

+++ ARM/include/app/demo.h
@@ -52,16 +52,16 @@
 /********************************************************************************************************/
 
 typedef struct {
-    uint8 hours;
-    uint8 minutes;
-    uint8 seconds;
-    uint8 dayPassed;
+    uint32 hours;
+    uint32 minutes;
+    uint32 seconds;
+    uint32 dayPassed;
 }CustomTime;
 
 typedef struct  {
-    uint8 day;
-    uint8 month;
-    uint8 year;
+    uint32 day;
+    uint32 month;
+    uint32 year;
 }CustomDate;
